refactor(frontend): decompose AutoTrading.tsx into AutoTrading/ directory
